refactor(scripts): split fixture generators into per-topic files
generate_fixtures.py was growing toward a god-file: shared infra,
two implemented generators, six stubs and the CLI shell all mixed
into 228 lines. Each new generator (watson, vgp, mcfadden,
fold_test, cutoff, bootstrap will land per Phase 1) added another
block to scroll past in review.

Layout now:
- _fixture_common.py — shared helpers, constants, StubNotImplemented
- gen_fisher.py      — gen_fisher
- gen_pca.py         — gen_pca (moved as-is; bug fix in next commit)
- generate_fixtures.py — argparse, GENERATORS dispatch, main (~95 LoC)

Flat in scripts/ rather than a fixture_generators/ subpackage —
filename matches CLI topic name (gen_pca.py ↔ `... pca`), no
__init__.py / package gymnastics. Stubs stay inline in the dispatch
table as 1-liners via gen_stub() — placeholder files would lie
about what's implemented.

Pure mechanical move. No logic changed; gen_pca's domean shape
bug is intentionally preserved here to keep the science fix
isolated and bisectable in the next commit.

Verified:
- ast.parse on all 4 files
- python3 scripts/generate_fixtures.py --help — argparse intact, 8 topics + all
- python3 scripts/generate_fixtures.py fisher — "no input files yet"
- python3 scripts/generate_fixtures.py fold_test — stub message, exit 1
- python3 scripts/generate_fixtures.py all — walks every topic, exit 0
